on this: go_move_current_only_to_root
new mutations controller almost complete
then I can test it
then I can refactor
then I can cut and paste to mutations with some added functions

thinking about a mutation.pre_mutation_id
to act as a fork to specify a mutation that links to possible clones of all its "pre" children,
and to offer an or-logic to normally and-logic mutation

if or logic is enabled,
The mutation will be able to offer a list of clone possibilities as a method of mutation completion.
When a clone possibility is selected, it is cloned as one of the children, so as to be one of the or logics selected.
Also, when the mutation itself is cloned, the previous mutations it is cloned from,
  all of their children will be offered as or-logic clone possibility.

    thinking about evolutions controller

    clone current yes mutations
    clone current no mutations
    clone current only yes mutations
    clone current only no mutations

    thinking about remote control

    new | clone | move | destroy
    current | current only
    no mutations | yes mutations